[Feature] Add payment (#21)
- Payment Dialog
- Store payment in order table
- Change table color/status when paid
- re-arrange files

+++ DAL/Order.cs
@@ -36,11 +36,11 @@ namespace DAL
         [Column(TypeName = "money")]
         public decimal? VAT { get; set; }
 
-        public int? Discount { get; set; }
+        public decimal? Discount { get; set; }
 
         public byte? DiscountType { get; set; }
 
-        public int? Extra { get; set; }
+        public decimal? Extra { get; set; }
 
         public String ExtraContent { get; set; }
